requirements.txt
tensorflow~=2.2.0
faiss-gpu==1.6.5
kapre==0.3.5
librosa
click
pyyaml
numpy
matplotlib
wavio
